fix(tblind): inject index properly
Previously, the inject_index API would take both a mutable reference to a partial signature and return a value, presumably to save some allocations.
This is error-prone and makes the interface counterintuitive. This commit fixes it to take non-mutable references and uses the returned value in the partial unblinding code