perf(states): do not clone status matrix
This commit uses interior mutability in the WaitingJustification struct to allow us to mutably borrow the StatusMatrix without
having to borrow the rest of the struct. This allows us to remove the redundant clone, which could be a problem in larger arrays.

We also get rid of the double hashmap lookup with a filter_map (previously it would do both `contains_key` and `get`)