Split DKG Core in modules (#48)
* refactor(dkg-core): split primitives in modules

* docs(states): expand docs and declare variables in order of use

* perf(states): remove redundant StatusMatrix clone when processing responses

The previous state is passed by value and not by reference. set_statuses should not allocate a new status matrix, rather it should just modify the existing one

* perf(states): do not clone status matrix

This commit uses interior mutability in the WaitingJustification struct to allow us to mutably borrow the StatusMatrix without
having to borrow the rest of the struct. This allows us to remove the redundant clone, which could be a problem in larger arrays.

We also get rid of the double hashmap lookup with a filter_map (previously it would do both `contains_key` and `get`)